pyproject.toml
[tool.poetry]
name = "image_classifier"
version = "0.1.0"
description = ""
authors = ["sarrrrry <sarrrrry.sarrrrry@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
hydra-core = "^0.11.3"
torch = "^1.5.0"
torchvision = "^0.6.0"
pandas = "^1.0.3"
tabulate = "^0.8.7"
mlflow = "^1.8.0"
flatten-dict = "^0.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
